requirements.txt
asgiref==3.2.10
boto==2.49.0
boto3==1.14.30
botocore==1.17.30
Django==3.0.8
django-filter==2.3.0
django-storages==1.9.1
docutils==0.15.2
gunicorn==20.0.4
jmespath==0.10.0
Pillow==7.2.0
python-dateutil==2.8.1
pytz==2020.1
s3transfer==0.3.3
six==1.15.0
sqlparse==0.3.1
urllib3==1.25.10
whitenoise==5.1.0
xlwt==1.3.0
